+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.3.2] - 2020-11-26
+### Added
+- `generateSSORequestParams` throws on empty PK
+
+### Changed
+- join operator for `messageToSign` is now `&`
+- `messageToSign` ignores data that are not set (`null` or `undefined`)
+
+### Fixed
+- ensure testing `throws` is working correctly
+
 ## [0.3.1] - 2020-11-25
 ### Added
 - Logo and badges to README